file broken into components
