+++ src/lab/Graphs/CalendarHeatmap/index.ts
@@ -1,2 +1,2 @@
-export type { WeekData } from "./base";
+export type { CalendarHeatmapTooltipProps, WeekData } from "./base";
 export * from "./CalendarHeatmap";
